requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.4
django-filter==23.3
urllib3==1.24.1
psycopg2==2.9.9
django-crontab==0.7.1